Chatbot_Final/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5
google-generativeai==0.8.3
sentence-transformers==3.3.1
faiss-cpu==1.9.0
pandas==2.2.3
numpy>=1.24.0,<2.0
pydantic==2.9.2
python-dotenv==1.0.1
pymongo==4.10.1